SDK-397: Extract an interface for Attribute<T>
