+++ yoti-sdk-api/src/main/java/com/yoti/api/client/Attribute.java
@@ -7,42 +7,13 @@ import java.util.Set;
  * {@link Profile} instances.
  *
  */
-public final class Attribute<T> {
+public interface Attribute<T> {
 
-    private final String name;
-    private final T value;
-    private final Set<String> sources;
-    private final Set<String> verifiers;
+    String getName();
 
-    public Attribute(String name, T value) {
-        this(name, value, null);
-    }
-    
-    public Attribute(String name, T value, Set<String> sources) {
-        this(name, value, sources, null);
-    }
+    T getValue();
 
-    public Attribute(String name, T value, Set<String> sources, Set<String> verifiers) {
-        this.name = name;
-        this.value = value;
-        this.sources = sources;
-        this.verifiers = verifiers;
-    }
+    Set<String> getSources();
 
-    public String getName() {
-        return name;
-    }
-
-    public T getValue() {
-        return value;
-    }
-
-    public Set<String> getSources() {
-        return sources;
-    }
-    
-    public Set<String> getVerifiers() {
-        return verifiers;
-    }
-
+    Set<String> getVerifiers();
 }
