+++ yoti-sdk-impl/src/test/java/com/yoti/api/client/spi/remote/SimpleProfileTest.java
@@ -125,7 +125,7 @@ public class SimpleProfileTest {
     }
 
     private static <T> List<Attribute<?>> asAttributeList(String key, T o) {
-        Attribute<?> a = new Attribute<>(key, o, null);
+        Attribute<?> a = new SimpleAttribute<>(key, o, null);
         return Collections.<Attribute<?>>singletonList(a);
     }
